refactor: move remaining composables to the category module
M2-555
