+++ packages/theme/composables/index.ts
@@ -10,8 +10,8 @@ export * from './useAddresses';
 export * from './useApi';
 export * from './useBilling';
 export * from './useCart';
-export * from './useCategory';
-export * from './useCategorySearch';
+export * from '~/modules/catalog/category/composables/useCategory';
+export * from '~/modules/catalog/category/composables/useCategorySearch';
 export * from './useConfig';
 export * from './useContent';
 export * from './useCountrySearch';
